gcs-trigger/requirements.txt
functions-framework==3.*
pandas
pandas_gbq
google-cloud-bigquery
google-cloud-storage
fsspec
gcsfs